+++ src/myuri/system.clj
@@ -1,6 +1,7 @@
 (ns myuri.system
   (:require [myuri.db :as db]
             [myuri.web.server :as server]
+            [myuri.link-checker :as checker]
             [com.stuartsierra.component :as component]))
 
 (defn new-system [config]
@@ -8,4 +9,6 @@
     (component/system-map
       :db (db/new-database database)
       :server (component/using (server/new-server server)
-                               [:db]))))
+                               [:db])
+      :linkchecker (component/using (checker/new-linkchecker)
+                                    [:db]))))