+++ src/myuri/web/templating.clj
@@ -3,7 +3,6 @@
             [ring.util.response :as resp]
             [selmer.parser :refer [render-file]]))
 
-
 (defn tpl-resp
   "docstring"
   ([template] (tpl-resp template nil))
@@ -11,6 +10,12 @@
    (resp/response
      {:selmer {:template template
                :data     data}})))
+
+(defn has-role?
+  "docstring"
+  [identity]
+  (fn [role]
+    (contains? (-> identity :roles) role)))
 
 (defn tpl-resp?
   "docstring"
@@ -24,8 +29,13 @@
     (let [res (handler req)]
       (if-let [selm (tpl-resp? res)]
         (let [{:keys [template data]} selm
-              tpl-data (merge {:app-addr (u/app-address req)
-                               :req      req}
+              tpl-data (merge {:app-addr   (u/app-address req)
+                               :req        req
+                               :identity (:identity req)
+                               :route-name (-> req :reitit.core/match :data :name)
+                               :has-role (has-role? (:identity req))
+                               }
+
                               data)]
           (assoc res :body (render-file template tpl-data)))
         res))))