+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Change Log
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
+## [0.11] – 2024-09-03
+### Added
+* User settings! Starting with control whether links should open in a new window or not.
+* Admin page with users overview
+
+### Changed
+* Search includes the bookmark description now
+* Introduced the is-admin flag for users; first user to register becomes admin
+
 ## [0.10] – 2024-08-24
 ### Fixed
 * Displayed username was wrong.